GroupProject4/DigitalZooManagementSystem/Django Backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary>=2.9
python-dateutil